IGNITE-28801 Calcite. Bump calcite dependency version to 1.42

+++ modules/calcite/src/test/java/org/apache/ignite/internal/processors/query/calcite/integration/FunctionsTest.java
@@ -86,7 +86,7 @@ public class FunctionsTest extends AbstractBasicIntegrationTest {
 
             assertQuery("SELECT BITAND(i, s) FROM TBL WHERE l=?").withParams(Short.MAX_VALUE).returns(0).check();
             assertQuery("SELECT BITOR(i, s) FROM TBL WHERE l=?").withParams(Short.MAX_VALUE).returns(65535).check();
-            assertQuery("SELECT BITXOR(i, s) FROM TBL WHERE l=?").withParams(Short.MAX_VALUE).returns(65535).check();
+            assertQuery("SELECT BITXOR(i, s) FROM TBL WHERE l=?").withParams(Short.MAX_VALUE).returns(65535L).check();
         }
         finally {
             sql("DROP TABLE if EXISTS TBL");
@@ -191,18 +191,18 @@ public class FunctionsTest extends AbstractBasicIntegrationTest {
         res.add(F.asList("XOR", 1.0, null, 1, null, xorErr));
         res.add(F.asList("XOR", 1, null, 1.0f, null, xorErr));
         res.add(F.asList("XOR", 1.0f, null, 1, null, xorErr));
-        res.add(F.asList("XOR", 1, null, 1, null, 0));
-        res.add(F.asList("XOR", 1, null, 0, null, 1));
-        res.add(F.asList("XOR", 0, null, 1, null, 1));
+        res.add(F.asList("XOR", 1, null, 1, null, 0L));
+        res.add(F.asList("XOR", 1, null, 0, null, 1L));
+        res.add(F.asList("XOR", 0, null, 1, null, 1L));
         res.add(F.asList("XOR", 1, null, 1, "BIGINT", 0L));
-        res.add(F.asList("XOR", 1, "TINYINT", 1, "INT", 0));
-        res.add(F.asList("XOR", 1, "TINYINT", 1, "SMALLINT", (short)0));
-        res.add(F.asList("XOR", 0, "TINYINT", 0, "TINYINT", (byte)0));
-        res.add(F.asList("XOR", 1, "TINYINT", 1, "SMALLINT", (short)0));
-        res.add(F.asList("XOR", 8, null, 7, null, 15));
-        res.add(F.asList("XOR", -1, null, 1, null, -2));
-        res.add(F.asList("XOR", (short)32767, null, 65535, null, 32768));
-        res.add(F.asList("XOR", (short)32767, null, 65536, null, 98303));
+        res.add(F.asList("XOR", 1, "TINYINT", 1, "INT", 0L));
+        res.add(F.asList("XOR", 1, "TINYINT", 1, "SMALLINT", 0L));
+        res.add(F.asList("XOR", 0, "TINYINT", 0, "TINYINT", 0L));
+        res.add(F.asList("XOR", 1, "TINYINT", 1, "SMALLINT", 0L));
+        res.add(F.asList("XOR", 8, null, 7, null, 15L));
+        res.add(F.asList("XOR", -1, null, 1, null, -2L));
+        res.add(F.asList("XOR", (short)32767, null, 65535, null, 32768L));
+        res.add(F.asList("XOR", (short)32767, null, 65536, null, 98303L));
         res.add(F.asList("XOR", null, null, 1, "TINYINT", null));
         res.add(F.asList("XOR", 1, null, null, null, null));
